Move GPG keys for EOL releases into another folder
rpm-ostree currently loads all the keys from the `/etc/pki/rpm-gpg/`
folder at startup to be able to verify ostree commits signatures.

This moves most of the now unused keys to a side folder
`/etc/pki/rpm-gpg.eol/` to keep them available but not in the default
path used by rpm-ostree.

This will break the repository definitions for older releases but those
are most likely not used anymore and fixing it is as easy as copying the
keys back to the non-eol folder for the required keys.

This will help all rpm-ostree based variants of Fedora (IoT, CoreOS,
Silverblue/Kinoite).

Just to be safe, we keep the keys for all the actively maintained
releases (F34 and later) as well as the last EOL release (F33).

See: https://github.com/coreos/fedora-coreos-tracker/issues/761