Create also rawhide+1, -1 and -2 key links
In separate software, it is useful to have also rawhide+1 key link prepared.
Especially when branching occurs and rawhide key is not yet updated, but
signing of packages has already started with a new key.

rawhide+1 and -1 are used in /etc/mock/templates/fedora-rawhide.tpl.
rawhide-1 and -2 are used in /etc/mock/templates/fedora-eln.tpl.